chore: Move protocols inside evm directory
The protocols implemented so far are evm specific so they should be inside this directory

--- don't change below this line ---
ENG-3786 Took 11 minutes


Took 2 minutes
