+++ src/evm/mod.rs
@@ -1,5 +1,6 @@
 pub mod account_storage;
 pub mod database;
+pub mod protocol;
 pub mod simulation;
 pub mod traces;
 pub mod tycho_db;
