+++ src/evm/protocol/state.rs
@@ -18,8 +18,8 @@
 //! # Examples
 //! ```
 //! use ethers::types::U256;
-//! use protosim::protocol::state::{ProtocolSim};
-//! use protosim::protocol::uniswap_v2::state::UniswapV2State;
+//! use protosim::evm::protocol::state::{ProtocolSim};
+//! use protosim::evm::protocol::uniswap_v2::state::UniswapV2State;
 //! use protosim::models::ERC20Token;
 //!
 //! // Initialize the UniswapV2 state with token reserves
@@ -47,12 +47,12 @@ use ethers::types::U256;
 use tycho_core::dto::ProtocolStateDelta;
 
 use crate::{
-    models::ERC20Token,
-    protocol::{
+    evm::protocol::{
         errors::{TradeSimulationError, TransitionError},
         events::{EVMLogMeta, LogIndex},
         models::GetAmountOutResult,
     },
+    models::ERC20Token,
 };
 
 /// ProtocolSim trait
