+++ src/evm/protocol/uniswap_v2/events.rs
@@ -2,7 +2,7 @@ use std::any::Any;
 
 use ethers::types::U256;
 
-use crate::protocol::state::ProtocolEvent;
+use crate::evm::protocol::state::ProtocolEvent;
 
 #[derive(Debug, Clone)]
 pub struct UniswapV2Sync {
