+++ src/evm/protocol/uniswap_v3/tick_math.rs
@@ -1,7 +1,7 @@
 use std::ops::BitOr;
 
 use crate::{
-    protocol::errors::TradeSimulationError,
+    evm::protocol::errors::TradeSimulationError,
     safe_math::{safe_div_u256, safe_mul_u256},
 };
 use ethers::types::{Sign, I256, U256};
