+++ src/evm/protocol/vm/assets/ISwapAdapter.abi
