+++ src/lib.rs
@@ -20,7 +20,6 @@ pub use num_traits;
 #[cfg(feature = "evm")]
 pub mod evm;
 pub mod models;
-pub mod protocol;
 pub mod safe_math;
 pub mod serde_helpers;
 #[cfg(feature = "evm")]
