+++ src/safe_math.rs
@@ -5,7 +5,7 @@
 //! will be returned.
 //! Functions for the types I256, U256, U512 are available.
 
-use crate::protocol::errors::{TradeSimulationError, TradeSimulationErrorKind};
+use crate::evm::protocol::errors::{TradeSimulationError, TradeSimulationErrorKind};
 use ethers::types::{I256, U256, U512};
 
 pub fn safe_mul_u256(a: U256, b: U256) -> Result<U256, TradeSimulationError> {
